DP-3715: Remove shortlinks causing broken links (#873)
The link tags with rel=shortlink in the head go to the internal
path in Drupal by default. Because the internal path (node/<nid>)
is being blocked in the .htaccess file, these are being identified
as broken links. Since shortlinks beyond this are not being
generated, these links are not necessary. This change removes
these links from all pages.