DP-4232 - Implement schema.org properties for "Service Detail Page"
Implemented schema.org data structure for the Service Details Page (content type). It is mapped to "GovernmentService" as detailed in http://schema.org/GovernmentService. Please note that not all fields in Drupal for this content type are mapped; however, work suggesting future mapping is documented in a sheet on the Google Drive. If applicable, candidate Drupal fields are also suggested.