Dp 4722 watch email link to diff (#1046)
* Refactor _send_notification for published nodes

Remove call to canonical_node because it is not behaving as
expected.  Originally, we wanted to confirm the node was
published before sending out any notifications because we
were calling this function for both hook_node_update and
hook_node_insert.

We now only call this function on update which takes away
some of the burden this was trying to solve.  Also, we determined
that canonical_node was actually not canonical and was the
node being sent from the hook.  This allowed for us to remove
it.

* Refactor watch email white list

It was determined that we were white listing the original
author of the node. This means that they would never get a
notification while all other watchers would.  To resolve this
we are now looking at the revision author.

* Refactor flag var names in send_notification

Var names were a little sloppy. Hopefully the following
convention gives more insight into what the code does.

* Revert "Refactor _send_notification for published nodes"

This reverts commit 11cb68028f93b11344880c0bd35933d017ba398c.

* Create diff_url for email

Had a hard time getting this to work.  What I ended up doing
was going to diff module and see how it was called.  In
PluginRevisionController.php, we see that a filter is passed.
This was the final step to get this Url working.

* Add diff_link to email

* t() will include lines breaks

* Revert "Refactor watch email white list"

This reverts commit ecbabefb7af0baf9ab58e53781c58f826f6edc3f.

* White list revision author from getting watching notifications

* Escape url links in email

* Revision url now correctly diffs old vs new revision

* Change formatting of method call to new line

* The person who performs the action will not receive a notification

* Remove queue_ui from lock

* Update text for watching emails

* Email output is now \r\n insteat of \n

* add node type to email and correct formatting

* remove formatting for vars and update text of email

* run composer update --lock